Move cache fetching methods into warm_cache, move warm_cache to automations.membership
